requirements.txt
django~=4.2.23
djangorestframework==3.16.1
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.1
pip~=25.2.0
pillow~=11.3.0
asgiref~=3.9.0
sqlparse~=0.5.3
psycopg2-binary==2.9.9
ruff==0.12.8
black==25.1.0
drf-nested-routers==0.94.2
pillow==11.3.0
